% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbiutils-package.R
\docType{package}
\name{dbiutils-package}
\alias{dbiutils}
\alias{dbiutils-package}
\title{dbiutils: Lightweight Utilities for DBI Workflows}
\description{
Small, driver-independent helpers for creating connections from YAML,
executing DBI queries, managing short-lived connections, and loading SQL
statements from files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ultreia-io.github.io/dbiutils/}
  \item \url{https://github.com/ultreia-io/dbiutils}
  \item Report bugs at \url{https://github.com/ultreia-io/dbiutils/issues}
}

}
\author{
\strong{Maintainer}: Tony Chemit \email{tchemit@users.noreply.github.com}

Authors:
\itemize{
  \item Tony Chemit \email{tchemit@users.noreply.github.com}
}

}
\keyword{internal}
